# Conflicts:
#   Java/Servicios/src/main/java/com/ekt/Servicios/controller/GroupController.java

+++ Java/Servicios/src/main/java/com/ekt/Servicios/controller/GroupController.java
@@ -5,9 +5,10 @@ import com.ekt.Servicios.entity.BodyUpdateBoss;
 import com.ekt.Servicios.entity.BodyAddUserGroup;
 import com.ekt.Servicios.entity.Group;
 import com.ekt.Servicios.entity.Response;
-import com.ekt.Servicios.entity.User;
 import com.ekt.Servicios.service.GroupService;
 import com.ekt.Servicios.service.UserService;
+
+
 import org.json.JSONObject;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,23 +36,66 @@ public class GroupController {
 
     @PutMapping("/saveUserGrup")
     public ResponseEntity<?> addUserGrup(@RequestBody BodyAddUserGroup bodyAddUserGroup){
-        return ResponseEntity.status(HttpStatus.CREATED).body(groupService.saveUser(bodyAddUserGroup.getIDUsuario(),bodyAddUserGroup.getIDGrupo(), bodyAddUserGroup.getIDSuperior(), bodyAddUserGroup.getNombreRol()));
+
+        if(bodyAddUserGroup.getIDUsuario()==null || bodyAddUserGroup.getIDGrupo()==null || bodyAddUserGroup.getIDSuperior()==null||bodyAddUserGroup.getNombreRol()==null){
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body(new Response(HttpStatus.NOT_ACCEPTABLE,"Error  en las llaves",""));
+        }else{
+            Group group= groupService.saveUser(bodyAddUserGroup.getIDUsuario(),bodyAddUserGroup.getIDGrupo(), bodyAddUserGroup.getIDSuperior(), bodyAddUserGroup.getNombreRol());
+            if (group==null){
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Response(HttpStatus.BAD_REQUEST,"Error al realizar en la operacion,parametro no valido",""));
+            }else{
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(new Response(HttpStatus.ACCEPTED,"El usuario se añadio correctamente",group));
+            }
+        }
     }
 
     @GetMapping("/buscar/{id}")
-    public Optional<Group> buscar(@PathVariable String id){
-        return groupService.findById(id);
+    public ResponseEntity<?> buscar(@PathVariable String id){
+
+        if (groupService.findById(id).isPresent()){
+            return ResponseEntity.ok(new Response(HttpStatus.ACCEPTED,"Grupo encontrado",groupService.findById(id)));
+        }else{
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Response(HttpStatus.BAD_REQUEST,"Error grupo no existente",""));
+        }
+
+
+        //return groupService.findById(id);
     }
 
     @DeleteMapping(value="/delete/{id}")
-    public void delete(@PathVariable String id){
-        System.out.println(id);
-        groupService.deleteById(id);
+    public ResponseEntity<Response> delete(@PathVariable String id){
+        if (id==null){
+            System.out.println("Error en las llaves");
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body(new Response(HttpStatus.NOT_ACCEPTABLE,"Error en las llaves",""));
+        }else{
+            Optional<Group> grupo= groupService.findById(id);
+            if (grupo.isPresent()){
+                groupService.deleteById(id);
+                System.out.println("Grupo eliminado");
+                return ResponseEntity.ok(new Response(HttpStatus.ACCEPTED,"Grupo eliminado",grupo.get()));
+            }else {
+                return ResponseEntity.ok(new Response(HttpStatus.BAD_REQUEST,"Grupo no encontrado",""));
+            }
+        }
     }
 
     @DeleteMapping("/deleteUserFromGruop")
-    public void deleteUserFromgroup(@RequestBody BodyAddUserGroup body){
-        groupService.deleteUserFromGroup(body.getIDUsuario(), body.getIDGrupo());
+    public ResponseEntity deleteUserFromgroup(@RequestBody BodyAddUserGroup body){
+        if (body.getIDGrupo()==null || body.getIDUsuario()==null){
+            System.out.println("Error en las llaves");
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body(new Response(HttpStatus.NOT_ACCEPTABLE,"Error en las llaves",""));
+        }else{
+            Optional<Group> grupo= groupService.userInGroup(body.getIDGrupo(),body.getIDUsuario());
+
+            if (grupo.isPresent()){
+                groupService.deleteUserFromGroup(body.getIDUsuario(), body.getIDGrupo());
+                System.out.println("Usuario eliminado del grupo");
+                return ResponseEntity.ok(new Response(HttpStatus.ACCEPTED,"Usuario eliminado del grupo",grupo.get()));
+            }else {
+                return ResponseEntity.ok(new Response(HttpStatus.BAD_REQUEST,"Grupo o usuario no encontrado",""));
+            }
+        }
+
     }
 
     @GetMapping("/findUserInGroup")
